# Conflicts:
#	modules/core/core-repo/src/main/java/com/enonic/xp/repo/impl/elasticsearch/query/ElasticsearchQuery.java
#	modules/core/core-repo/src/main/java/com/enonic/xp/repo/impl/elasticsearch/snapshot/AbstractSnapshotExecutor.java
#	modules/core/core-repo/src/main/java/com/enonic/xp/repo/impl/elasticsearch/snapshot/SnapshotExecutor.java
#	modules/core/core-repo/src/main/java/com/enonic/xp/repo/impl/elasticsearch/snapshot/SnapshotRestoreExecutor.java
#	modules/core/core-repo/src/main/java/com/enonic/xp/repo/impl/elasticsearch/suggistion/query/SuggestionQueryBuilderFactory.java

+++ modules/core/core-repo/src/main/java/com/enonic/xp/repo/impl/elasticsearch/result/SearchExplanationFactory.java
@@ -24,7 +24,7 @@ class SearchExplanationFactory
         final QueryExplanation.Builder builder = QueryExplanation.create();
 
         builder.description( explanation.getDescription() );
-        builder.value( explanation.getValue() );
+        builder.value( explanation.getValue().floatValue() );
 
         if ( explanation.getDetails() != null )
         {
